[echo] Isolate repository indexing
Separate the rolling GitHub file index from MarinMirror ingestion and centralize federated-search configuration. Remove unused index fields and keep behavior tests coupled to endpoint contracts rather than terminal rendering.
