+++ infra/echo/api/Dockerfile
@@ -13,7 +13,7 @@ ENV PYTHONUNBUFFERED=1
 RUN pip install --no-cache-dir \
     "fastapi>=0.115" "uvicorn>=0.30" \
     "cloud-sql-python-connector[pg8000]>=1.9" "sqlalchemy>=2" "pgvector>=0.3" \
-    "fastembed>=0.8"
+    "fastembed>=0.8,<0.9"
 
 WORKDIR /app
 COPY schema.py .
@@ -24,6 +24,6 @@ COPY api/dashboard.py .
 COPY --from=dashboard /build/dist /app/dist
 
 # Bake the query embedding model into the image so the first request is not a cold download.
-RUN python -c "from fastembed import TextEmbedding; TextEmbedding('BAAI/bge-small-en-v1.5')"
+RUN python -c "import search_config; from fastembed import TextEmbedding; TextEmbedding(search_config.EMBED_MODEL)"
 
 CMD ["sh", "-c", "exec uvicorn app:app --host 0.0.0.0 --port ${PORT:-8080}"]
