+++ infra/echo/search_config.py
@@ -4,6 +4,7 @@
 """Search settings shared by Echo's schema and query statements."""
 
 EMBED_MODEL = "BAAI/bge-small-en-v1.5"
+SEARCH_DOMAINS = ("wiki", "file", "discord", "pr", "issue")
 TEXT_SEARCH_CONFIG = "english"
 TS_RANK_NORMALIZATION = 32
 DEFAULT_SEARCH_LIMIT = 10
